Load plugin on all of okcupid. Otherwise I can't reliably detect being on the /questions page.
See https://stackoverflow.com/questions/20865581/chrome-extension-content-script-not-loaded-until-page-is-refreshed

Basically, I have the red borders right now and can easily see when my code is loaded. If I come to a /questions page cleanly from a new tab or a full refresh, it loads. But if I come from the home page and then a profile, it doesn't. The SO page has no better suggestions than expanding the scope of the extension. I'm not happy with that but good enough for now I guess.

Performance is a concern. I'm waiting for the question list to exist. And now sometimes it never will if you're on the wrong page. The ugly compromise for now is to check less often. But it's an infinite loop :(

Possible idea, not being explored yet: https://stackoverflow.com/a/39508954/1541186
Might be able to detect when the link changes myself. (I just happened to still have this tab up from the MutationObserver stuff, and it might actually be relevant.)