Completely rework loading on the page.
The previoius approach didn't work. OKCupid does too much DOM manipulation. I couldn't trust the element I was waiting on would still exist, and so I'd miss the mutations I wanted.

This goes back to the old approach, but a bit more elegantly, and with promises. It listens for the correct items. Not just the elements on the page, but also the absence of any loading info.

Also does some cleanup. Much (all?) from the previous commit removed, and some from before then. I think everything left here is important.

A bit of restructuring to make future dev easier.