Save CSV data to the firefox storage api
The file is now only a fallback. I'll phase it out.

This mostly works in the debugger. But the promises getting stuff from the backend are failing sometimes and not logging what the failure is. Needs a bit more tweaking, but this is close.
Amended: WIth a clean reload, this seems to be working great!