First pass at adding counts.
Shows a range from the number we know will be visible to the number that could be including those off screen.
Does not include uncategorized, and probably should.
Also doesn't know the upper bound on filter_id pages that aren't the main three as there's no OKCupid element to read.