Add an Edit Filter button. Refactor logic per question type, including removing the borders.
It's theoretically possible for this to have been doable without the logic refactor, but it's cleaner now anyway. Keeping the distinct question type methods around would have been good only for the border colors, and I was on the fence on removing those anyway. I just got rid of them, so this will solve two issues.