Also listen for updates to the questions section of the page.
I'm not sure how this previously worked when scrolling down and loading more questions. It must have triggered a change to the filters somehow, and no longer does?

I'm not sure the list is needed, but it might be with adding/removing filters, and that is how it was originally written, so I'll leave it.