Only rework the whole page on new URL if the profile changed
A recent OKCupid change is updating the URL whenever the user switches between the agree/disagree/findout filters. It didn't do this in the past. This was causing us to think we were on a brand new question page each time, and we'd keep adding more add/edit/delete filter buttons from the manipulatePage call.